refactor: split remaining behavior suites
